+++ golang/testsuite/testsuite_impl/example_testsuite.go
@@ -15,14 +15,23 @@ import (
 	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl/network_partition_test"
 )
 
+/*
+	NEW USER ONBOARDING:
+	- Refactor the name of the myCustomServiceImage property to reflect the name of your service.
+*/
 type ExampleTestsuite struct {
+	myCustomServiceImage string
 	apiServiceImage string
 	datastoreServiceImage string
 	isKurtosisCoreDevMode bool
 }
 
-func NewExampleTestsuite(apiServiceImage string, datastoreServiceImage string, isKurtosisCoreDevMode bool) *ExampleTestsuite {
-	return &ExampleTestsuite{apiServiceImage: apiServiceImage, datastoreServiceImage: datastoreServiceImage, isKurtosisCoreDevMode: isKurtosisCoreDevMode}
+/*
+	NEW USER ONBOARDING:
+	- Refactor the name of the myCustomServiceImage argument to reflect the name of your service.
+*/
+func NewExampleTestsuite(myCustomServiceImage string, apiServiceImage string, datastoreServiceImage string, isKurtosisCoreDevMode bool) *ExampleTestsuite {
+	return &ExampleTestsuite{myCustomServiceImage: myCustomServiceImage, apiServiceImage: apiServiceImage, datastoreServiceImage: datastoreServiceImage, isKurtosisCoreDevMode: isKurtosisCoreDevMode}
 }
 
 func (suite ExampleTestsuite) GetTests() map[string]testsuite.Test {